# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Weekly API
  version: 0.6.0
  description: 주간보고 작성·복제, Confluence 6.9.1 자동 초안, AI 구조화, PPTX Import, 관리자 설정, 분석 및 키 관리 REST API
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /version:
    get:
      security: []
      summary: 빌드 버전 조회
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /auth/providers:
    get:
      security: []
      summary: 로그인 방식과 공개 서비스 정보
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /auth/login:
    post:
      security: []
      summary: 로컬 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
  /me:
    get:
      summary: 현재 사용자와 서비스 컨텍스트
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /reports/current:
    get:
      summary: 현재 주차 본인 보고서
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /reports/current/candidates:
    get:
      summary: 현재 주차 Confluence 자동 초안과 Sync 상태
      responses:
        "200":
          description: 사용자별 자동 초안
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Envelope" }
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          enabled: { type: boolean }
                          weekStart: { type: string, format: date }
                          candidates:
                            type: array
                            items: { $ref: "#/components/schemas/ConfluenceCandidate" }
  /report-candidates/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    patch:
      summary: 본인 Confluence 자동 초안 수정 및 사용자 수정 보호
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfluenceCandidateWrite" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      summary: 본인 자동 초안을 제외하고 같은 주차 재생성 방지
      responses:
        "204": { description: 제외 완료 }
        "404": { $ref: "#/components/responses/Error" }
  /report-candidates/{id}/sources:
    get:
      summary: 본인 자동 초안의 Confluence 원본 Page 목록
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
  /report-candidates/accept:
    post:
      summary: 자동 초안을 저장한 본인 보고서에 연결
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, reportId]
              properties:
                ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: integer, format: int64 }
                reportId: { type: integer, format: int64 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /reports:
    get:
      summary: 권한 범위 보고서 검색
      parameters:
        - { in: query, name: weekStart, schema: { type: string, format: date } }
        - { in: query, name: status, schema: { $ref: "#/components/schemas/ReportStatus" } }
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: 보고서 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportWrite" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
  /reports/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      summary: 보고서 상세
      responses:
        "200": { $ref: "#/components/responses/Success" }
    put:
      summary: 본인 보고서 낙관적 잠금 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportWrite" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      summary: 본인 보고서 낙관적 잠금 삭제
      parameters:
        - in: query
          name: version
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: 삭제 완료 }
        "409": { $ref: "#/components/responses/Error" }
  /reports/{id}/submit:
    post:
      summary: 보고서 제출 또는 즉시 확정
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /reports/{id}/clone:
    post:
      summary: 본인 보고서를 새 주차의 독립된 초안으로 복제
      description: 승인 이력, 댓글, 외부 출처 연결과 원본 항목 ID는 복사하지 않는다. 대상 주차가 없으면 현재 주차를 사용한다.
      parameters: [{ $ref: "#/components/parameters/ID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetWeekStart: { type: string, format: date, description: 관리자가 설정한 주차 시작 요일이어야 함 }
                mode: { type: string, enum: [STRUCTURE, FULL], default: STRUCTURE }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /reports/{id}/approve:
    post:
      summary: 제출 보고서 승인
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /reports/{id}/reject:
    post:
      summary: 제출 보고서 반려
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /reports/{id}/export.pptx:
    get:
      summary: 템플릿 기반 PPTX 다운로드
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200":
          description: PowerPoint 파일
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema: { type: string, format: binary }
  /ai/reports/parse-text:
    post:
      summary: 자유 텍스트를 저장하지 않고 주간보고 미리보기로 구조화
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 200000 }
      responses:
        "200":
          description: 검증된 AI 미리보기
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Envelope" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AIWeeklyResult" }
        "503": { $ref: "#/components/responses/Error" }
  /import/pptx:
    post:
      summary: 과거 PPTX 다중 업로드 및 비동기 분석 작업 생성
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "202": { $ref: "#/components/responses/Success" }
        "503": { $ref: "#/components/responses/Error" }
  /import/history:
    get:
      summary: 본인의 PPTX Import 작업 이력
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /import/{id}:
    get:
      summary: 본인의 Import 작업과 파일별 분석 미리보기
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200":
          description: Import 작업 상세
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Envelope" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ImportJob" }
        "404": { $ref: "#/components/responses/Error" }
  /import/{id}/analyze:
    post:
      summary: 보관 원본이 있는 실패 파일 재분석
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "202": { $ref: "#/components/responses/Success" }
  /import/{id}/confirm:
    post:
      summary: 사용자가 수정한 과거 보고 미리보기 확정
      parameters: [{ $ref: "#/components/parameters/ID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: "#/components/schemas/ImportConfirmFile" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /keys:
    get:
      summary: 개인 API 키 목록
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: 개인 API 키 발급
      responses:
        "201": { $ref: "#/components/responses/Success" }
  /keys/rotate:
    post:
      summary: 개인 키 버전 회전 및 전부 폐기
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /rollups:
    get:
      summary: 월간·분기·반기·연간 기간 보고 집계
      description: |
        기간에 겹치는 주간보고를 취합해 중복 업무와 반복 기재를 제거한 업무 목록,
        완료율·정체·이슈 지속·보고 커버리지 인사이트와 주차별 추이를 반환합니다.
        저장하지 않고 요청할 때마다 원본 주간보고에서 다시 계산합니다.
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [MONTH, QUARTER, HALF, YEAR], default: MONTH } }
        - { name: period, in: query, description: "2026-08, 2026-Q3, 2026-H2, 2026. 생략하면 현재 기간", schema: { type: string } }
        - { name: scope, in: query, description: "TEAM은 팀장 이상만 조회할 수 있습니다.", schema: { type: string, enum: [SELF, TEAM], default: SELF } }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /rollups/export.csv:
    get:
      summary: 기간 보고 업무 표 CSV 내보내기
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [MONTH, QUARTER, HALF, YEAR], default: MONTH } }
        - { name: period, in: query, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string, enum: [SELF, TEAM], default: SELF } }
      responses:
        "200":
          description: UTF-8 BOM CSV
          content:
            text/csv:
              schema: { type: string, format: binary }
  /rollups/export.pptx:
    get:
      summary: 기간 보고 PPTX 내보내기
      description: 주간보고와 동일한 관리자 등록 PPTX 규격으로 기간 집계 결과를 내보냅니다.
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [MONTH, QUARTER, HALF, YEAR], default: MONTH } }
        - { name: period, in: query, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string, enum: [SELF, TEAM], default: SELF } }
      responses:
        "200":
          description: PPTX 파일
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema: { type: string, format: binary }
        "409": { $ref: "#/components/responses/Error" }
  /analytics/overview:
    get:
      summary: 조직 보고 현황 분석
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /analytics/endpoints:
    get:
      summary: 최근 24시간 API 분석, 관리자 전용
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/settings:
    get:
      summary: 운영 설정 조회
      responses:
        "200": { $ref: "#/components/responses/Success" }
    put:
      summary: 운영 설정 변경
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/settings/ai/test:
    post:
      summary: 저장된 AI Gateway 설정의 Structured Output 연결 시험
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "502": { $ref: "#/components/responses/Error" }
  /admin/settings/confluence/test:
    post:
      summary: 저장된 Confluence 6.9.1 REST 설정 연결 시험
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "502": { $ref: "#/components/responses/Error" }
  /admin/confluence/sync:
    post:
      summary: 관리자 증분 동기화 비동기 요청
      responses:
        "202": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/confluence/sync/status:
    get:
      summary: 동기화 상태·처리 건수·최근 오류
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/confluence/users/mappings:
    get:
      summary: Weekly와 Confluence 사용자 매핑·자동 추천 목록
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/confluence/users/unmapped:
    get:
      summary: 활성 미매핑 사용자 목록
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/confluence/users/{id}/mapping:
    put:
      summary: 관리자 명시 Confluence 사용자 매핑
      parameters:
        - { $ref: "#/components/parameters/ID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUsername, active]
              properties:
                externalUsername: { type: string, minLength: 1, maxLength: 255 }
                active: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/users:
    get:
      summary: 사용자 목록
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: 사용자 등록
      responses:
        "201": { $ref: "#/components/responses/Success" }
  /admin/organizations:
    get:
      summary: 조직 목록
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: 조직 등록
      responses:
        "201": { $ref: "#/components/responses/Success" }
  /admin/pptx-template:
    get:
      summary: PPTX 템플릿 정보
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: PPTX 템플릿 업로드
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200": { $ref: "#/components/responses/Success" }
    delete:
      summary: 기본 PPTX 템플릿 복원
      responses:
        "200": { $ref: "#/components/responses/Success" }
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: weekly_session }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    ID: { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
  responses:
    Success:
      description: 성공
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
  schemas:
    Envelope:
      type: object
      required: [success, traceId]
      properties:
        success: { type: boolean }
        data: {}
        error:
          type: [object, "null"]
          properties:
            code: { type: string }
            message: { type: string }
        traceId: { type: string }
    ReportStatus:
      type: string
      enum: [DRAFT, SUBMITTED, REVISION_REQUESTED, APPROVED, CLOSED]
    ReportItem:
      type: object
      required: [title, currentResult, nextPlan, issue, progress]
      properties:
        category: { type: string }
        title: { type: string, maxLength: 240 }
        currentResult: { type: string }
        nextPlan: { type: string }
        issue: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
    AIReportItem:
      allOf:
        - { $ref: "#/components/schemas/ReportItem" }
        - type: object
          required: [confidence]
          properties:
            confidence: { type: number, minimum: 0, maximum: 1 }
    AIWeeklyResult:
      type: object
      required: [summary, weekStart, dateConfidence, reportItems, warnings]
      properties:
        summary: { type: string }
        weekStart: { type: string, description: 명확하지 않으면 빈 문자열 }
        dateConfidence: { type: number, minimum: 0, maximum: 1 }
        reportItems:
          type: array
          items: { $ref: "#/components/schemas/AIReportItem" }
        warnings:
          type: array
          items: { type: string }
    ImportFile:
      type: object
      required: [id, originalFilename, fileHash, sizeBytes, status]
      properties:
        id: { type: integer, format: int64 }
        originalFilename: { type: string }
        fileHash: { type: string, pattern: "^[0-9a-f]{64}$" }
        sizeBytes: { type: integer, format: int64 }
        status:
          type: string
          enum: [QUEUED, PROCESSING, READY, NEEDS_REVIEW, FAILED, DUPLICATE, CONFIRMED, SKIPPED]
        detectedWeekStart: { type: string }
        detectedWeekEnd: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        dateSource: { type: string }
        result: { $ref: "#/components/schemas/AIWeeklyResult" }
        errorMessage: { type: string }
        duplicateOf: { type: integer, format: int64 }
        reportId: { type: integer, format: int64 }
        conflictReportId: { type: integer, format: int64 }
    ImportJob:
      type: object
      required: [id, status, totalFiles, processedFiles, failedFiles, createdAt]
      properties:
        id: { type: integer, format: int64 }
        status:
          type: string
          enum: [PENDING, PROCESSING, READY, PARTIAL, FAILED, CONFIRMED]
        totalFiles: { type: integer }
        processedFiles: { type: integer }
        failedFiles: { type: integer }
        createdAt: { type: string, format: date-time }
        files:
          type: array
          items: { $ref: "#/components/schemas/ImportFile" }
    ImportConfirmFile:
      type: object
      required: [id, weekStart, strategy, items]
      properties:
        id: { type: integer, format: int64 }
        weekStart: { type: string, format: date }
        summary: { type: string }
        strategy: { type: string, enum: [CREATE, MERGE, REPLACE, SKIP] }
        items:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: "#/components/schemas/AIReportItem" }
    ConfluenceSource:
      type: object
      required: [pageId, title, spaceKey, pageUrl, pageVersion, activityType]
      properties:
        pageId: { type: string }
        title: { type: string }
        spaceKey: { type: string }
        pageUrl: { type: string, format: uri }
        pageVersion: { type: integer }
        activityType: { type: string, enum: [CREATED, MODIFIED, CREATED_AND_MODIFIED] }
        sourceUpdatedAt: { type: string, format: date-time }
    ConfluenceCandidateWrite:
      type: object
      required: [normalizedTitle, category, currentResult, nextPlan, issue]
      properties:
        normalizedTitle: { type: string, minLength: 1, maxLength: 240 }
        category: { type: string, maxLength: 80 }
        currentResult: { type: string }
        nextPlan: { type: string }
        issue: { type: string }
    ConfluenceCandidate:
      allOf:
        - { $ref: "#/components/schemas/ConfluenceCandidateWrite" }
        - type: object
          required: [id, weekStart, confidence, ruleScore, status, userEdited, sources]
          properties:
            id: { type: integer, format: int64 }
            weekStart: { type: string, format: date }
            confidence: { type: number, minimum: 0, maximum: 1 }
            ruleScore: { type: integer }
            status: { type: string, enum: [DETECTED, ACCEPTED, IGNORED, MERGED, REMOVED] }
            userEdited: { type: boolean }
            sources:
              type: array
              items: { $ref: "#/components/schemas/ConfluenceSource" }
    ReportWrite:
      type: object
      required: [summary, items]
      properties:
        weekStart: { type: string, format: date }
        summary: { type: string, maxLength: 10000 }
        version: { type: integer, minimum: 1 }
        sourceType: { type: string, enum: [MANUAL, AI_TEXT, CONFLUENCE_AI] }
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ReportItem" }